BIOS functionality - add receiving bios info func

+++ sys/gpu/gpu.go
@@ -2,7 +2,7 @@ package gpu
 
 type GPUInfo struct {
 	GPUName        string `json:"name"`
-	MemoryBytes    uint64 `json:"memory"`
+	MemoryBytes    uint32 `json:"memory"`
 	DriverVersion  string `json:"driverVersion"`
 	VideoProcessor string `json:"videoProcessor"`
 }
